+++ pkg/resources/resource_monitor.go
@@ -1,14 +1,12 @@
 package resources
 
 import (
+	"context"
 	"database/sql"
-	"errors"
 	"fmt"
-	"log"
-	"strconv"
-	"strings"
 
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/snowflake"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/helpers"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
@@ -130,6 +128,7 @@ func checkAccountAgainstWarehouses(d *schema.ResourceData, name string) error {
 // CreateResourceMonitor implements schema.CreateFunc.
 func CreateResourceMonitor(d *schema.ResourceData, meta interface{}) error {
 	db := meta.(*sql.DB)
+	client := sdk.NewClientFromDB(db)
 	name := d.Get("name").(string)
 
 	check := checkAccountAgainstWarehouses(d, name)
@@ -138,65 +137,88 @@ func CreateResourceMonitor(d *schema.ResourceData, meta interface{}) error {
 		return check
 	}
 
-	cb := snowflake.NewResourceMonitorBuilder(name).Create()
+	ctx := context.Background()
+	objectIdentifier := sdk.NewAccountObjectIdentifier(name)
+
 	// Set optionals.
+	opts := &sdk.CreateResourceMonitorOptions{}
 	if v, ok := d.GetOk("notify_users"); ok {
-		cb.SetStringList("notify_users", expandStringList(v.(*schema.Set).List()))
+		userNames := expandStringList(v.(*schema.Set).List())
+		users := []sdk.NotifiedUser{}
+		for _, name := range userNames {
+			users = append(users, sdk.NotifiedUser{Name: name})
+		}
+		if opts.With == nil {
+			opts.With = &sdk.ResourceMonitorWith{}
+		}
+		opts.With.NotifyUsers = &sdk.NotifyUsers{Users: users}
 	}
+
 	if v, ok := d.GetOk("credit_quota"); ok {
-		cb.SetInt("credit_quota", v.(int))
+		if opts.With == nil {
+			opts.With = &sdk.ResourceMonitorWith{}
+		}
+		opts.With.CreditQuota = sdk.Int(v.(int))
 	}
 	if v, ok := d.GetOk("frequency"); ok {
-		cb.SetString("frequency", v.(string))
+		frequency, err := sdk.FrequencyFromString(v.(string))
+		if err != nil {
+			return err
+		}
+		if opts.With == nil {
+			opts.With = &sdk.ResourceMonitorWith{}
+		}
+		opts.With.Frequency = frequency
 	}
+
 	if v, ok := d.GetOk("start_timestamp"); ok {
-		cb.SetString("start_timestamp", v.(string))
+		if opts.With == nil {
+			opts.With = &sdk.ResourceMonitorWith{}
+		}
+		opts.With.StartTimestamp = sdk.Pointer(v.(string))
 	}
 	if v, ok := d.GetOk("end_timestamp"); ok {
-		cb.SetString("end_timestamp", v.(string))
-	}
-	if v, ok := d.GetOk("suspend_trigger"); ok {
-		cb.SuspendAt(v.(int))
-	}
-	// Support deprecated suspend_triggers.
-	if v, ok := d.GetOk("suspend_triggers"); ok {
-		siTrigs := expandIntList(v.(*schema.Set).List())
-		for _, t := range siTrigs {
-			cb.SuspendAt(t)
+		if opts.With == nil {
+			opts.With = &sdk.ResourceMonitorWith{}
 		}
-	}
-	if v, ok := d.GetOk("suspend_immediate_trigger"); ok {
-		cb.SuspendImmediatelyAt(v.(int))
-	}
-	// Support deprecated suspend_immediate_triggers.
-	if v, ok := d.GetOk("suspend_immediate_triggers"); ok {
-		sTrigs := expandIntList(v.(*schema.Set).List())
-		for _, t := range sTrigs {
-			cb.SuspendImmediatelyAt(t)
-		}
-	}
-	nTrigs := expandIntList(d.Get("notify_triggers").(*schema.Set).List())
-	for _, t := range nTrigs {
-		cb.NotifyAt(t)
+		opts.With.EndTimestamp = sdk.Pointer(v.(string))
 	}
 
-	stmt := cb.Statement()
-	if err := snowflake.Exec(db, stmt); err != nil {
+	triggers := collectResourceMonitorTriggers(d)
+	if len(triggers) > 0 {
+		if opts.With == nil {
+			opts.With = &sdk.ResourceMonitorWith{}
+		}
+		opts.With.Triggers = triggers
+	}
+
+	err := client.ResourceMonitors.Create(ctx, objectIdentifier, opts)
+	if err != nil {
 		return fmt.Errorf("error creating resource monitor %v err = %w", name, err)
 	}
-
 	d.SetId(name)
 
 	if d.Get("set_for_account").(bool) {
-		if err := snowflake.Exec(db, cb.SetOnAccount()); err != nil {
+		accountOpts := sdk.AlterAccountOptions{
+			Set: &sdk.AccountSet{
+				ResourceMonitor: objectIdentifier,
+			},
+		}
+		if err := client.Accounts.Alter(ctx, &accountOpts); err != nil {
 			return fmt.Errorf("error setting resource monitor %v on account err = %w", name, err)
 		}
 	}
 
 	if v, ok := d.GetOk("warehouses"); ok {
 		for _, w := range v.(*schema.Set).List() {
-			if err := snowflake.Exec(db, cb.SetOnWarehouse(w.(string))); err != nil {
-				return fmt.Errorf("error setting resource monitor %v on warehouse %v err = %w", name, w.(string), err)
+			warehouseOpts := sdk.AlterWarehouseOptions{
+				Set: &sdk.WarehouseSet{
+					ResourceMonitor: objectIdentifier,
+				},
+			}
+			warehouseId := sdk.NewAccountObjectIdentifier(w.(string))
+			if err := client.Warehouses.Alter(ctx, warehouseId, &warehouseOpts); err != nil {
+				return fmt.Errorf("error setting resource monitor %v on warehouse %v err = %w", name, warehouseId.Name(), err)
 			}
 		}
 	}
@@ -210,134 +232,68 @@ func CreateResourceMonitor(d *schema.ResourceData, meta interface{}) error {
 // ReadResourceMonitor implements schema.ReadFunc.
 func ReadResourceMonitor(d *schema.ResourceData, meta interface{}) error {
 	db := meta.(*sql.DB)
-	stmt := snowflake.NewResourceMonitorBuilder(d.Id()).Show()
+	client := sdk.NewClientFromDB(db)
+	objectIdentifier := helpers.DecodeSnowflakeID(d.Id()).(sdk.AccountObjectIdentifier)
 
-	row := snowflake.QueryRow(db, stmt)
-
-	rm, err := snowflake.ScanResourceMonitor(row)
-	if errors.Is(err, sql.ErrNoRows) {
-		// If not found, mark resource to be removed from state file during apply or refresh
-		log.Printf("[DEBUG] resource monitor (%s) not found", d.Id())
-		d.SetId("")
-		return nil
-	}
+	ctx := context.Background()
+	resourceMonitor, err := client.ResourceMonitors.ShowByID(ctx, objectIdentifier)
 	if err != nil {
 		return err
 	}
 
-	// Set string values
-	nullStrings := map[string]sql.NullString{
-		"name":            rm.Name,
-		"frequency":       rm.Frequency,
-		"start_timestamp": rm.StartTime,
-		"end_timestamp":   rm.EndTime,
+	if err := d.Set("name", resourceMonitor.Name); err != nil {
+		return err
 	}
-	if err := setDataFromNullStrings(d, nullStrings); err != nil {
+	if err := d.Set("frequency", string(resourceMonitor.Frequency)); err != nil {
 		return err
 	}
 
-	if len(rm.NotifyUsers.String) > 0 {
-		if err := d.Set("notify_users", strings.Split(rm.NotifyUsers.String, ", ")); err != nil {
+	if err := d.Set("start_timestamp", resourceMonitor.StartTime); err != nil {
+		return err
+	}
+
+	if err := d.Set("end_timestamp", resourceMonitor.EndTime); err != nil {
+		return err
+	}
+
+	if len(resourceMonitor.NotifyUsers) > 0 {
+		if err := d.Set("notify_users", resourceMonitor.NotifyUsers); err != nil {
 			return err
 		}
 	}
 
 	// Snowflake returns credit_quota as a float, but only accepts input as an int
-	if rm.CreditQuota.Valid {
-		cqf, err := strconv.ParseFloat(rm.CreditQuota.String, 64)
-		if err != nil {
-			return err
-		}
-		if err := d.Set("credit_quota", int(cqf)); err != nil {
-			return err
-		}
+	if err := d.Set("credit_quota", int(resourceMonitor.CreditQuota)); err != nil {
+		return err
 	}
 
 	// Triggers
-	sTrig, err := extractTriggerInts(rm.SuspendAt)
-	if err != nil {
-		return err
-	}
-
-	var setSuspendTrigger bool
-	if _, ok := d.GetOk("suspend_trigger"); ok {
-		if len(sTrig) > 0 {
-			if err := d.Set("suspend_trigger", sTrig[0]); err != nil {
-				return err
-			}
-			setSuspendTrigger = true
-		} else {
-			if err := d.Set("suspend_trigger", nil); err != nil {
-				return err
-			}
-			setSuspendTrigger = true
-		}
-	}
-	if _, ok := d.GetOk("suspend_triggers"); ok {
-		if err := d.Set("suspend_triggers", sTrig); err != nil {
+	if resourceMonitor.SuspendAt != nil {
+		if err := d.Set("suspend_trigger", *resourceMonitor.SuspendAt); err != nil {
 			return err
 		}
-		setSuspendTrigger = true
-	}
-	if !setSuspendTrigger {
-		if len(sTrig) > 0 {
-			if err := d.Set("suspend_trigger", sTrig[0]); err != nil {
-				return err
-			}
-		} else {
-			if err := d.Set("suspend_trigger", nil); err != nil {
-				return err
-			}
-		}
-	}
-
-	siTrig, err := extractTriggerInts(rm.SuspendImmediatelyAt)
-	if err != nil {
-		return err
-	}
-
-	var setSuspendImmediateTrigger bool
-	if _, ok := d.GetOk("suspend_immediate_trigger"); ok {
-		if len(siTrig) > 0 {
-			if err := d.Set("suspend_immediate_trigger", siTrig[0]); err != nil {
-				return err
-			}
-			setSuspendImmediateTrigger = true
-		} else {
-			if err := d.Set("suspend_immediate_trigger", nil); err != nil {
-				return err
-			}
-			setSuspendImmediateTrigger = true
-		}
-	}
-	if _, ok := d.GetOk("suspend_immediate_triggers"); ok {
-		if err := d.Set("suspend_immediate_triggers", siTrig); err != nil {
+	} else {
+		if err := d.Set("suspend_trigger", nil); err != nil {
 			return err
 		}
-		setSuspendImmediateTrigger = true
 	}
-	if !setSuspendImmediateTrigger {
-		if len(siTrig) > 0 {
-			if err := d.Set("suspend_immediate_trigger", siTrig[0]); err != nil {
-				return err
-			}
-		} else {
-			if err := d.Set("suspend_immediate_trigger", nil); err != nil {
-				return err
-			}
+
+	if resourceMonitor.SuspendImmediateAt != nil {
+		if err := d.Set("suspend_immediate_trigger", *resourceMonitor.SuspendImmediateAt); err != nil {
+			return err
+		}
+	} else {
+		if err := d.Set("suspend_immediate_trigger", nil); err != nil {
+			return err
 		}
 	}
 
-	nTrigs, err := extractTriggerInts(rm.NotifyAt)
-	if err != nil {
-		return err
-	}
-	if err := d.Set("notify_triggers", nTrigs); err != nil {
+	if err := d.Set("notify_triggers", resourceMonitor.NotifyTriggers); err != nil {
 		return err
 	}
 
 	// Account level
-	if err := d.Set("set_for_account", rm.Level.Valid && rm.Level.String == "ACCOUNT"); err != nil {
+	if err := d.Set("set_for_account", resourceMonitor.Level == sdk.ResourceMonitorLevelAccount); err != nil {
 		return err
 	}
 
@@ -360,135 +316,131 @@ func setDataFromNullStrings(data *schema.ResourceData, ns map[string]sql.NullStr
 	return nil
 }
 
-// extractTriggerInts converts the triggers in the DB (stored as a comma
-// separated string with trailing %s) into a slice of ints.
-func extractTriggerInts(s sql.NullString) ([]int, error) {
-	// Check if this is NULL
-	if !s.Valid {
-		return []int{}, nil
-	}
-	ints := strings.Split(s.String, ",")
-	out := make([]int, 0, len(ints))
-	for _, i := range ints {
-		myInt, err := strconv.Atoi(i[:len(i)-1])
-		if err != nil {
-			return out, fmt.Errorf("failed to convert %v to integer err = %w", i, err)
-		}
-		out = append(out, myInt)
-	}
-	return out, nil
-}
-
 // UpdateResourceMonitor implements schema.UpdateFunc.
 func UpdateResourceMonitor(d *schema.ResourceData, meta interface{}) error {
 	db := meta.(*sql.DB)
-	id := d.Id()
+	client := sdk.NewClientFromDB(db)
+	name := d.Get("name").(string)
 
-	check := checkAccountAgainstWarehouses(d, id)
+	check := checkAccountAgainstWarehouses(d, name)
 
 	if check != nil {
 		return check
 	}
 
-	ub := snowflake.NewResourceMonitorBuilder(id).Alter()
+	objectIdentifier := helpers.DecodeSnowflakeID(d.Id()).(sdk.AccountObjectIdentifier)
+
+	ctx := context.Background()
 	var runSetStatement bool
+
+	opts := sdk.AlterResourceMonitorOptions{Set: &sdk.ResourceMonitorSet{}}
 
 	if d.HasChange("notify_users") {
 		runSetStatement = true
-		ub.SetStringList(`NOTIFY_USERS`, expandStringList(d.Get("notify_users").(*schema.Set).List()))
+
+		userNames := expandStringList(d.Get("notify_users").(*schema.Set).List())
+		users := []sdk.NotifiedUser{}
+		for _, name := range userNames {
+			users = append(users, sdk.NotifiedUser{Name: name})
+		}
+		opts.NotifyUsers.Users = users
 	}
 
 	if d.HasChange("credit_quota") {
 		runSetStatement = true
-		ub.SetInt(`CREDIT_QUOTA`, d.Get("credit_quota").(int))
+		opts.Set.CreditQuota = sdk.Pointer(d.Get("credit_quota").(int))
 	}
 
 	if d.HasChange("frequency") {
 		runSetStatement = true
-		ub.SetString(`FREQUENCY`, d.Get("frequency").(string))
+		frequency, err := sdk.FrequencyFromString(d.Get("frequency").(string))
+		if err != nil {
+			return err
+		}
+		opts.Set.Frequency = frequency
 	}
 
 	if d.HasChange("start_timestamp") {
 		runSetStatement = true
-		ub.SetString(`START_TIMESTAMP`, d.Get("start_timestamp").(string))
+		opts.Set.StartTimestamp = sdk.Pointer(d.Get("start_timestamp").(string))
 	}
 
 	if d.HasChange("end_timestamp") {
 		runSetStatement = true
-		ub.SetString(`END_TIMESTAMP`, d.Get("end_timestamp").(string))
+		opts.Set.EndTimestamp = sdk.Pointer(d.Get("end_timestamp").(string))
 	}
 
-	// Set triggers
-	if d.HasChange("suspend_trigger") {
+	// If ANY of the triggers changed, we collect all triggers and set them
+	if d.HasChange("suspend_trigger") ||
+		d.HasChange("suspend_triggers") ||
+		d.HasChange("suspend_immediate_trigger") ||
+		d.HasChange("suspend_immediate_triggers") ||
+		d.HasChange("notify_triggers") {
 		runSetStatement = true
-		ub.SuspendAt(d.Get("suspend_trigger").(int))
-	}
-	// Support deprecated suspend_triggers.
-	if d.HasChange("suspend_triggers") {
-		runSetStatement = true
-		siTrigs := expandIntList(d.Get("suspend_triggers").(*schema.Set).List())
-		for _, t := range siTrigs {
-			ub.SuspendAt(t)
-		}
-	}
-
-	if d.HasChange("suspend_immediate_trigger") {
-		runSetStatement = true
-		ub.SuspendImmediatelyAt(d.Get("suspend_immediate_trigger").(int))
-	}
-	// Support deprecated suspend_immediate_trigger.
-	if d.HasChange("suspend_immediate_triggers") {
-		runSetStatement = true
-		siTrigs := expandIntList(d.Get("suspend_immediate_triggers").(*schema.Set).List())
-		for _, t := range siTrigs {
-			ub.SuspendImmediatelyAt(t)
-		}
-	}
-
-	nTrigs := expandIntList(d.Get("notify_triggers").(*schema.Set).List())
-	for _, t := range nTrigs {
-		runSetStatement = true
-		ub.NotifyAt(t)
+		triggers := collectResourceMonitorTriggers(d)
+		opts.Triggers = triggers
 	}
 
 	if runSetStatement {
-		if err := snowflake.Exec(db, ub.Statement()); err != nil {
-			return fmt.Errorf("error updating resource monitor %v\n%w", id, err)
+		if err := client.ResourceMonitors.Alter(ctx, objectIdentifier, &opts); err != nil {
+			return fmt.Errorf("error updating resource monitor %v\n%w", objectIdentifier.Name(), err)
 		}
 	}
 
 	// Remove from account
 	if d.HasChange("set_for_account") && !d.Get("set_for_account").(bool) {
-		if err := snowflake.Exec(db, ub.UnsetOnAccount()); err != nil {
-			return fmt.Errorf("error unsetting resource monitor %v on account err = %w", id, err)
+		accountOpts := sdk.AlterAccountOptions{
+			Set: &sdk.AccountSet{
+				ResourceMonitor: sdk.NewAccountObjectIdentifier("NULL"),
+			},
+		}
+		if err := client.Accounts.Alter(ctx, &accountOpts); err != nil {
+			return fmt.Errorf("error unsetting resource monitor %v on account err = %w", objectIdentifier.Name(), err)
 		}
 	}
 
 	// Remove from all old warehouses
 	if d.HasChange("warehouses") {
 		oldV, v := d.GetChange("warehouses")
-		res := intersectionAAndNotB(oldV.(*schema.Set).List(), v.(*schema.Set).List())
+		res := ADiffB(oldV.(*schema.Set).List(), v.(*schema.Set).List())
 		for _, w := range res {
-			if err := snowflake.Exec(db, ub.UnsetOnWarehouse(w)); err != nil {
-				return fmt.Errorf("error setting resource monitor %v on warehouse %v err = %w", id, w, err)
+			warehouseOpts := sdk.AlterWarehouseOptions{
+				Unset: &sdk.WarehouseUnset{
+					ResourceMonitor: sdk.Bool(true),
+				},
+			}
+			warehouseId := sdk.NewAccountObjectIdentifier(w)
+			if err := client.Warehouses.Alter(ctx, warehouseId, &warehouseOpts); err != nil {
+				return fmt.Errorf("error unsetting resource monitor %v on warehouse %v err = %w", name, warehouseId.Name(), err)
 			}
 		}
 	}
 
 	// Add to account
 	if d.HasChange("set_for_account") && d.Get("set_for_account").(bool) {
-		if err := snowflake.Exec(db, ub.SetOnAccount()); err != nil {
-			return fmt.Errorf("error setting resource monitor %v on account err = %w", id, err)
+		accountOpts := sdk.AlterAccountOptions{
+			Set: &sdk.AccountSet{
+				ResourceMonitor: objectIdentifier,
+			},
+		}
+		if err := client.Accounts.Alter(ctx, &accountOpts); err != nil {
+			return fmt.Errorf("error setting resource monitor %v on account err = %w", name, err)
 		}
 	}
 
 	// Add to all new warehouses
 	if d.HasChange("warehouses") {
 		oldV, v := d.GetChange("warehouses")
-		res := intersectionAAndNotB(v.(*schema.Set).List(), oldV.(*schema.Set).List())
+		res := ADiffB(v.(*schema.Set).List(), oldV.(*schema.Set).List())
 		for _, w := range res {
-			if err := snowflake.Exec(db, ub.SetOnWarehouse(w)); err != nil {
-				return fmt.Errorf("error setting resource monitor %v on warehouse %v err = %w", id, w, err)
+			warehouseOpts := sdk.AlterWarehouseOptions{
+				Set: &sdk.WarehouseSet{
+					ResourceMonitor: objectIdentifier,
+				},
+			}
+			warehouseId := sdk.NewAccountObjectIdentifier(w)
+			if err := client.Warehouses.Alter(ctx, warehouseId, &warehouseOpts); err != nil {
+				return fmt.Errorf("error setting resource monitor %v on warehouse %v err = %w", name, warehouseId.Name(), err)
 			}
 		}
 	}
@@ -496,13 +448,74 @@ func UpdateResourceMonitor(d *schema.ResourceData, meta interface{}) error {
 	return ReadResourceMonitor(d, meta)
 }
 
+func collectResourceMonitorTriggers(d *schema.ResourceData) []sdk.TriggerDefinition {
+	triggers := []sdk.TriggerDefinition{}
+	var suspendTrigger *sdk.TriggerDefinition
+	if v, ok := d.GetOk("suspend_trigger"); ok {
+		suspendTrigger = &sdk.TriggerDefinition{
+			Threshold:     v.(int),
+			TriggerAction: sdk.TriggerActionSuspend,
+		}
+	}
+
+	if v, ok := d.GetOk("suspend_triggers"); ok {
+		siTrigs := expandIntList(v.(*schema.Set).List())
+		for _, threshold := range siTrigs {
+			if suspendTrigger == nil || suspendTrigger.Threshold > threshold {
+				suspendTrigger = &sdk.TriggerDefinition{
+					Threshold:     threshold,
+					TriggerAction: sdk.TriggerActionSuspend,
+				}
+			}
+		}
+	}
+	if suspendTrigger != nil {
+		triggers = append(triggers, *suspendTrigger)
+	}
+	var suspendImmediateTrigger *sdk.TriggerDefinition
+
+	if v, ok := d.GetOk("suspend_immediate_trigger"); ok {
+		suspendImmediateTrigger = &sdk.TriggerDefinition{
+			Threshold:     v.(int),
+			TriggerAction: sdk.TriggerActionSuspendImmediate,
+		}
+	}
+
+	if v, ok := d.GetOk("suspend_immediate_triggers"); ok {
+		siTrigs := expandIntList(v.(*schema.Set).List())
+		for _, threshold := range siTrigs {
+			if suspendImmediateTrigger == nil || suspendTrigger.Threshold > threshold {
+				suspendImmediateTrigger = &sdk.TriggerDefinition{
+					Threshold:     threshold,
+					TriggerAction: sdk.TriggerActionSuspendImmediate,
+				}
+			}
+		}
+	}
+	if suspendImmediateTrigger != nil {
+		triggers = append(triggers, *suspendImmediateTrigger)
+	}
+
+	nTrigs := expandIntList(d.Get("notify_triggers").(*schema.Set).List())
+	for _, t := range nTrigs {
+		triggers = append(triggers, sdk.TriggerDefinition{
+			Threshold:     t,
+			TriggerAction: sdk.TriggerActionNotify,
+		})
+	}
+	return triggers
+}
+
 // DeleteResourceMonitor implements schema.DeleteFunc.
 func DeleteResourceMonitor(d *schema.ResourceData, meta interface{}) error {
 	db := meta.(*sql.DB)
+	client := sdk.NewClientFromDB(db)
+	ctx := context.Background()
+	objectIdentifier := helpers.DecodeSnowflakeID(d.Id()).(sdk.AccountObjectIdentifier)
 
-	stmt := snowflake.NewResourceMonitorBuilder(d.Id()).Drop()
-	if err := snowflake.Exec(db, stmt); err != nil {
-		return fmt.Errorf("error deleting resource monitor %v err = %w", d.Id(), err)
+	err := client.ResourceMonitors.Drop(ctx, objectIdentifier)
+	if err != nil {
+		return err
 	}
 
 	d.SetId("")